feat: move starlette application back to entry point
